refactor: modularize directory structure
