+++ src/app/AppReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers, Reducer } from 'redux';
-import { counterReducer } from '../counter';
-import { globalReducer } from '../global';
+import { counterReducer } from '../pages/counter';
+import { globalReducer } from '../routing';
 import { AppAction, AppState } from './app-type';
 
 export const appReducer: Reducer<AppState, AppAction> = combineReducers<
