+++ src/app/useRedux.ts
@@ -1,7 +1,7 @@
 import { useContext } from 'react';
 // @ts-ignore
 import { ReactReduxContext } from 'react-redux';
-import { AppDispatch, AppState } from './app/app-type';
+import { AppDispatch, AppState } from './app-type';
 
 export function useRedux<T>(
   mapState: (state: AppState) => T,
