+++ src/pages/counter/index.ts
@@ -1,1 +1,3 @@
 export { Counter } from './Counter';
+export { counterEffects } from './CounterEffect';
+export { counterReducer } from './CounterReducer';
