+++ src/registerAll.ts
@@ -1,0 +1,5 @@
+import 'firebase/app';
+import 'firebase/auth';
+import 'firebase/firestore';
+import 'firebaseui/dist/firebaseui.css';
+import './registerFontAwesome';
