+++ src/routing/utils/getCurrentPath.ts
