--- src/store/app/index.ts
@@ -1,1 +1,0 @@
-export { appReducer } from './AppReducer';
