--- src/store/counter/index.ts
@@ -1,2 +1,0 @@
-export { counterEffects } from './CounterEffect';
-export { counterReducer } from './CounterReducer';
